requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
openpyxl>=3.1.0
opencv-python>=4.8.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=9.5.0
numpy>=1.24.0
